Split utils into examine and math.
